requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.2
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
chess==1.9.0
chess.com==1.7.5
click==7.1.2
coc==0.2
config==0.5.1
cycler==0.10.0
decorator==5.0.9
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
ghostscript==0.7
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.3.1
logomaker==0.8
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.1
packaging==20.9
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.3
seqlogo==5.29.8
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.18
toml==0.10.2
torch==1.11.0
traitlets==5.0.5
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
weblogo==3.7.8
Werkzeug==1.0.1
yarl==1.6.3
